feat: multi-host architecture with embedded node agent
Implements issue #33 - multi-host control plane + node agents architecture
with embedded node support for single-host deployments.

## Changes:

### Crate Reorganization
- Renamed minions-host → minions-vsock-cli (clarifies it's a VSOCK test tool)
- Created minions-db crate (shared database types and operations)
- Created minions-node crate (library + standalone binary)
  - Extracted vm, hypervisor, network, storage, agent modules from minions
  - Provides both in-process API (for embedded mode) and HTTP API (for remote)

### Database Schema
- Added `hosts` table for tracking node agents (id, address, capacity, status)
- Added `host_id` column to `vms` table
- Created default 'local' host for embedded node agent (single-host deployments)

### Control Plane (minions crate)
- New `scheduler` module: chooses host for VM placement based on capacity
- New `host_client` module: HTTP client for calling remote node agents
- Updated `vm` module: dispatches to local (minions_node::*) vs remote (host_client)
  - Local host: calls minions_node library directly (in-process, zero overhead)
  - Remote host: makes HTTP call to node agent API

### Single-Host Deployment (Embedded Mode)
- Default behavior unchanged - single-process, all-in-one
- No WireGuard, no HTTP overhead, no extra services
- Node logic runs in-process via minions_node library

### Multi-Host Deployment (Optional)
- Standalone `minions-node` binary runs on remote hosts (requires --features standalone)
- Control plane makes HTTP calls to remote agents
- Scheduler chooses host based on available CPU/memory

## Not Yet Implemented (follow-up work):
- Host management CLI (`minions host add/list/remove`)
- Background tasks (heartbeat checker, capacity sync)
- WireGuard overlay networking (required for multi-host)
- Fixes for standalone minions-node binary (has compilation issues)
- Remote snapshot/resize/copy operations (currently local-only)
- Tests for scheduler and multi-host dispatch
