+++ crates/minions-db/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "minions-db"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+rusqlite = { version = "0.32", features = ["bundled"] }
+chrono = "0.4"
+uuid = { version = "1", features = ["v4"] }
