+++ crates/minions-vsock-cli/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "minions-host"
+name = "minions-vsock-cli"
 version = "0.1.0"
 edition = "2024"
 
 [[bin]]
-name = "minions-host"
+name = "minions-vsock-cli"
 path = "src/main.rs"
 
 [dependencies]
