+++ crates/minions/Cargo.toml
@@ -13,6 +13,12 @@ path = "../minions-ssh"
 [dependencies.minions-proxy]
 path = "../minions-proxy"
 
+[dependencies.minions-node]
+path = "../minions-node"
+
+[dependencies.minions-db]
+path = "../minions-db"
+
 [dependencies]
 askama = "0.12"
 minions-proto = { path = "../minions-proto" }
